- some Integration service tests that uses multithreading fails when they run with other tests, but they succeed alone.
- probably multithreading problem, or the machine can't serve enough threads for the test before timeouts. timeouts were increased but still no success till now.

+++ src/main/java/com/nasnav/service/ProductImageServiceImpl.java
@@ -3,12 +3,20 @@ package com.nasnav.service;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_CSV_PARSE_FAILURE;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_IMPORTING_IMGS;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_IMPORTING_IMG_FILE;
+import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_NO_IMG_DATA_PROVIDED;
+import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_NO_IMG_IMPORT_RESPONSE;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_NO_PRODUCT_EXISTS_WITH_BARCODE;
+import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_NO_PRODUCT_EXISTS_WITH_ID;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_PRODUCT_IMG_BULK_IMPORT;
 import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_READ_ZIP;
+import static com.nasnav.constatnts.error.dataimport.ErrorMessages.ERR_USER_CANNOT_MODIFY_PRODUCT;
+import static java.lang.String.format;
+import static java.util.Comparator.comparing;
+import static java.util.stream.Collectors.groupingBy;
 
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -37,13 +45,16 @@ import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
+import com.nasnav.commons.utils.EntityUtils;
 import com.nasnav.constatnts.EntityConstants.Operation;
 import com.nasnav.dao.EmployeeUserRepository;
+import com.nasnav.dao.FilesRepository;
 import com.nasnav.dao.ProductImagesRepository;
 import com.nasnav.dao.ProductRepository;
 import com.nasnav.dao.ProductVariantsRepository;
 import com.nasnav.dto.ProductImageBulkUpdateDTO;
 import com.nasnav.dto.ProductImageUpdateDTO;
+import com.nasnav.dto.ProductImgDetailsDTO;
 import com.nasnav.exceptions.BusinessException;
 import com.nasnav.persistence.BaseUserEntity;
 import com.nasnav.persistence.ProductEntity;
@@ -87,7 +98,8 @@ public class ProductImageServiceImpl implements ProductImageService {
 	@Autowired
 	private SecurityService securityService;
 	
-	
+	@Autowired
+	private FilesRepository fileRepo;
 	
 	
 
@@ -97,7 +109,10 @@ public class ProductImageServiceImpl implements ProductImageService {
 		
 		return saveProductImg(file, imgMetaData);
 	}
-
+	
+	
+	
+	
 
 
 	/**
@@ -171,13 +186,57 @@ public class ProductImageServiceImpl implements ProductImageService {
 
 	private ProductImageUpdateResponse saveNewProductImg(MultipartFile file, ProductImageUpdateDTO imgMetaData)
 			throws BusinessException {
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		BaseUserEntity user =  empRepo.getOneByEmail(auth.getName());
+		List<ProductImageUpdateDTO> metaDataList = Arrays.asList(imgMetaData);
 		
-		String url = fileService.saveFile(file, user.getOrganizationId());
+		return saveNewProductImgsUsingSameUrl(file, metaDataList)
+					.stream()
+					.findFirst()
+					.orElseThrow(() -> new BusinessException(ERR_NO_IMG_IMPORT_RESPONSE, ERR_NO_IMG_IMPORT_RESPONSE, HttpStatus.INTERNAL_SERVER_ERROR));
+	}
+	
+	
+	
+	
+	
+	
+	private List<ProductImageUpdateResponse> saveNewProductImgsUsingSameUrl(MultipartFile file, List<ProductImageUpdateDTO> imgMetaDataList)
+			throws BusinessException {
+		if(imgMetaDataList == null || imgMetaDataList.isEmpty()) {
+			throw new BusinessException(ERR_NO_IMG_DATA_PROVIDED, ERR_NO_IMG_DATA_PROVIDED, HttpStatus.NOT_ACCEPTABLE);
+		}
 		
-		Long imgId = saveProductImgToDB(imgMetaData, url);
+		for(ProductImageUpdateDTO metaData: imgMetaDataList) {
+			validateProductImg(file, metaData);
+		}
 		
+		String url = saveFile(file);
+		
+		List<ProductImageUpdateResponse> responses = new ArrayList<>();
+		for(ProductImageUpdateDTO metaData: imgMetaDataList) {
+			responses.add( saveNewProductImgMetaData(metaData, url) );
+		}
+		
+		return responses;
+	}
+	
+
+
+	
+	
+
+	private String saveFile(MultipartFile file) throws BusinessException {
+		Long userOrgId =  securityService.getCurrentUserOrganizationId();		
+		return fileService.saveFile(file, userOrgId);
+	}
+	
+	
+	
+	
+	
+	
+	private ProductImageUpdateResponse saveNewProductImgMetaData(ProductImageUpdateDTO imgMetaData, String url)
+			throws BusinessException {		
+		Long imgId = saveProductImgToDB(imgMetaData, url);		
 		return new ProductImageUpdateResponse(imgId, url);
 	}
 
@@ -320,16 +379,27 @@ public class ProductImageServiceImpl implements ProductImageService {
 	 * product must follow the organization of the user 
 	 * */
 	private void validateUserCanModifyProduct(Optional<ProductEntity> product) throws BusinessException {
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		BaseUserEntity user =  empRepo.getOneByEmail(auth.getName());
-		Long userOrg = user.getOrganizationId();
-		Long productOrg = product.get().getOrganizationId();
+		BaseUserEntity user = securityService.getCurrentUser();
+		Long productOrg = product.map(p -> p.getOrganizationId()).orElse(null);
 		
-		if(!Objects.equals(userOrg, productOrg))
+		if(!Objects.equals( user.getOrganizationId() , productOrg))
 			throw new BusinessException(
-					String.format("User with email [%s] have no rights to modify products from organization of id[%d]!", user.getEmail(), productOrg)
+					format(ERR_USER_CANNOT_MODIFY_PRODUCT, user.getEmail(), productOrg)
 					, "INSUFFICIENT RIGHTS"
 					, HttpStatus.FORBIDDEN);
+	}
+	
+	
+	
+	
+	
+	
+	/**
+	 * product must follow the organization of the user 
+	 * */
+	private void validateUserCanModifyProduct(Long productId) throws BusinessException {
+		Optional<ProductEntity> product = productRepository.findById(productId);
+		validateUserCanModifyProduct(product);
 	}
 
 
@@ -376,11 +446,23 @@ public class ProductImageServiceImpl implements ProductImageService {
 		
 		validateImgToDelete(img);		
 		
+		Long cnt = productImagesRepository.countByUri(img.getUri());
 		productImagesRepository.deleteById(imgId);
 		
-		fileService.deleteFileByUrl(img.getUri());
+		deleteImgFileIfNotUsed(img,cnt);		
 		
 		return new ProductImageDeleteResponse(productId);
+	}
+
+
+
+
+
+
+	private void deleteImgFileIfNotUsed(ProductImagesEntity img, Long cnt) throws BusinessException {
+		if(cnt <= 1) {
+			fileService.deleteFileByUrl(img.getUri());
+		}
 	}
 
 
@@ -423,19 +505,14 @@ public class ProductImageServiceImpl implements ProductImageService {
 
 
 	private List<ProductImageUpdateResponse> saveImgsBulk(@Valid MultipartFile zip, @Valid MultipartFile csv,
-			@Valid ProductImageBulkUpdateDTO metaData) throws BusinessException {
-		List<ImportedImage> importedImgs = extractImgsToImport(zip, csv, metaData);
+			@Valid ProductImageBulkUpdateDTO metaData) throws BusinessException {		
 		List<String> errors = new ArrayList<>();
-		List<ProductImageUpdateResponse> responses = new ArrayList<>(); 
+		List<ProductImageUpdateResponse> responses = new ArrayList<>();
 		
-		for(ImportedImage img : importedImgs) {
-			try {
-				responses.add( updateProductImage(img.getImage(), img.getImgMetaData()) );
-			}catch(Exception e) {
-				logger.log(Level.SEVERE, e.getMessage(), e);				
-				
-				errors.add( createImgImportErrMsg(img, e) );
-			}
+		Map<String, List<ImportedImage>> imgsGroupedByFile = getFileToImgsMetatDataMap(zip, csv, metaData);
+		
+		for(String fileName : imgsGroupedByFile.keySet()) {
+			saveSingleImgToAllItsVariants(fileName, imgsGroupedByFile, errors, responses);
 		}
 		
 		if(!errors.isEmpty()) {
@@ -444,6 +521,66 @@ public class ProductImageServiceImpl implements ProductImageService {
 		}
 		
 		return responses;
+	}
+
+
+
+
+
+
+	private Map<String, List<ImportedImage>> getFileToImgsMetatDataMap(MultipartFile zip, MultipartFile csv,
+			ProductImageBulkUpdateDTO metaData) throws BusinessException {
+		List<ImportedImage> allImportedImgs = extractImgsToImport(zip, csv, metaData);
+		Map<String, List<ImportedImage>> groupedByFile = allImportedImgs.stream()
+																	.collect(Collectors.groupingBy(ImportedImage::getZipFileName));
+		return groupedByFile;
+	}
+
+
+
+
+
+
+	private void saveSingleImgToAllItsVariants(String fileName, Map<String, List<ImportedImage>> fileImgMetadataMap,
+			List<String> errors, List<ProductImageUpdateResponse> responses) {
+		MultipartFile file = getMulitpartFile(fileImgMetadataMap, fileName);			
+		List<ProductImageUpdateDTO> imgMetaDataList = getImgsMetaDataList(fileImgMetadataMap, fileName);
+		
+		try {
+			responses.addAll( saveNewProductImgsUsingSameUrl(file, imgMetaDataList) );
+		}catch(Exception e) {
+			logger.log(Level.SEVERE, e.getMessage(), e);						
+			fileImgMetadataMap.get(fileName)
+						.forEach(img -> errors.add( createImgImportErrMsg(img, e) ));
+		}
+	}
+
+
+
+
+
+
+	private List<ProductImageUpdateDTO> getImgsMetaDataList(Map<String, List<ImportedImage>> groupedByFile,
+			String fileName) {
+		List<ProductImageUpdateDTO> imgMetaDataList = groupedByFile.get(fileName)
+																	.stream()
+																	.map(ImportedImage::getImgMetaData)
+																	.collect(Collectors.toList());
+		return imgMetaDataList;
+	}
+
+
+
+
+
+
+	private MultipartFile getMulitpartFile(Map<String, List<ImportedImage>> groupedByFile, String fileName) {
+		MultipartFile file = groupedByFile.get(fileName)
+										.stream()
+										.map(ImportedImage::getImage)
+										.findFirst()
+										.orElse(null);
+		return file;
 	}
 
 
@@ -492,7 +629,7 @@ public class ProductImageServiceImpl implements ProductImageService {
 			@Valid ProductImageBulkUpdateDTO metaData) throws BusinessException {
 		List<ImportedImage> imgs = new ArrayList<>();		
 		List<String> errors = new ArrayList<>();		
-		Map<String,String> fileBarcodeMap = createFileBarcodeMap(csv);
+		Map<String,List<String>> fileBarcodeMap = createFileBarcodeMap(csv);
 		
 		try(ZipInputStream stream = new ZipInputStream(zip.getInputStream())){	
 			
@@ -518,7 +655,7 @@ public class ProductImageServiceImpl implements ProductImageService {
 	
 
 	private List<ImportedImage> readZipStream(ZipInputStream stream, ProductImageBulkUpdateDTO metaData,
-			Map<String, String> fileBarcodeMap, List<String> errors) throws IOException {
+			Map<String, List<String>> fileBarcodeMap, List<String> errors) throws IOException {
 		
 		List<ImportedImage> imgs = new ArrayList<>();
 		
@@ -547,7 +684,7 @@ public class ProductImageServiceImpl implements ProductImageService {
 	 * as a single barcode can belong to both a product and a product variant.
 	 * */
 	private List<ImportedImage> readImgsFromZipEntry(ZipEntry zipEntry, ZipInputStream stream,
-			ProductImageBulkUpdateDTO metaData, Map<String, String> fileBarcodeMap, List<String> errors) {
+			ProductImageBulkUpdateDTO metaData, Map<String, List<String>> fileBarcodeMap, List<String> errors) {
 		List<ImportedImage> imgsFromEntry = new ArrayList<>();
 		
 		if(zipEntry.isDirectory() ) {
@@ -559,7 +696,7 @@ public class ProductImageServiceImpl implements ProductImageService {
 			List<ProductImageUpdateDTO> imgsMetaData = createImportedImagesMetaData(zipEntry, fileBarcodeMap, metaData);				  
 			
 			imgsFromEntry = imgsMetaData.stream()
-									.map( meta -> new ImportedImage(imgMultipartFile, meta))
+									.map( meta -> new ImportedImage(imgMultipartFile, meta, zipEntry.getName()))
 									.collect(Collectors.toList());
 									
 		}catch(Exception e) {
@@ -606,9 +743,24 @@ public class ProductImageServiceImpl implements ProductImageService {
 	 * - barcode exists for variant only
 	 * - barcode exists for both a product and a variant
 	 * */
-	private List<ProductImageUpdateDTO> createImportedImagesMetaData(ZipEntry zipEntry,Map<String,String> fileBarcodeMap, ProductImageBulkUpdateDTO metaData) throws BusinessException{
+	private List<ProductImageUpdateDTO> createImportedImagesMetaData(ZipEntry zipEntry,Map<String,List<String>> fileBarcodeMap, ProductImageBulkUpdateDTO metaData) throws BusinessException{
 		
-		String barcode = getBarcodeOfImportedImg(zipEntry, fileBarcodeMap);		
+		List<String> barcodes = getBarcodeOfImportedImg(zipEntry, fileBarcodeMap);
+		
+		List<List<ProductImageUpdateDTO>> metaDataLists = new ArrayList<>();
+		for(String barcode: barcodes) {
+			metaDataLists.add(createImportedImagesMetaData(metaData, barcode));
+		}
+		
+		return metaDataLists.stream()
+					.flatMap(List::stream)
+					.collect(Collectors.toList());
+	}
+
+
+
+	private List<ProductImageUpdateDTO> createImportedImagesMetaData(ProductImageBulkUpdateDTO metaData, String barcode)
+			throws BusinessException {
 		Long orgId = securityService.getCurrentUserOrganizationId();
 		
 		
@@ -640,13 +792,15 @@ public class ProductImageServiceImpl implements ProductImageService {
 	
 
 
-	private String getBarcodeOfImportedImg(ZipEntry zipEntry, Map<String, String> fileBarcodeMap) {
+	private List<String> getBarcodeOfImportedImg(ZipEntry zipEntry, Map<String, List<String>> fileBarcodeMap) {
 		String fileName = zipEntry.getName();
-		String barcode = fileBarcodeMap.get(fileName);
-		if(barcode == null) {
-			barcode =  getBarcodeFromImgName(fileName);
+		List<String> barcodes = fileBarcodeMap.get(fileName);
+		
+		if(barcodes == null) {
+			String barcode =  getBarcodeFromImgName(fileName);
+			barcodes = Arrays.asList( barcode);
 		}
-		return barcode;
+		return barcodes;
 	}
 	
 	
@@ -726,7 +880,7 @@ public class ProductImageServiceImpl implements ProductImageService {
 
 	
 
-	private Map<String, String> createFileBarcodeMap(MultipartFile csv) throws BusinessException {
+	private Map<String, List<String>> createFileBarcodeMap(MultipartFile csv) throws BusinessException {
 		if(csv == null ||csv.isEmpty())
 			return new HashMap<>();		
 		
@@ -734,7 +888,8 @@ public class ProductImageServiceImpl implements ProductImageService {
 					.stream()
 					.collect(Collectors.toMap(
 											rec -> normalizeZipPath( rec.getString(1) )
-											, rec ->rec.getString(0)));
+											, rec -> Arrays.asList( rec.getString(0))
+											, EntityUtils::concateLists)); // resolve duplicated paths in the csv
 	}
 
 
@@ -816,10 +971,51 @@ public class ProductImageServiceImpl implements ProductImageService {
 							.stream()
 							.filter(Objects::nonNull)
 							.filter(this::isProductCoverImage)
-							.sorted(this::compareByProductImageId)
+							.sorted( comparing(ProductImagesEntity::getId))
 							.findFirst()
 							.map(img-> img.getUri())
 							.orElse(NO_IMG_FOUND_URL);
+	}
+	
+	
+	
+	
+	@Override
+	public Map<Long,String> getProductsCoverImages(List<Long> productIds) {
+		if(productIds == null || productIds.isEmpty()) {
+			return new HashMap<>();
+		}
+		
+		Map<Long,String> productImgs = productImagesRepository
+											.findByProductEntity_IdInOrderByPriority(productIds)
+											.stream()
+											.filter(Objects::nonNull)
+											.filter(this::isProductCoverImage)
+											.collect( groupingBy(img -> img.getProductEntity().getId()))
+											.entrySet()
+											.stream()
+											.map(this::getProductCoverImageUrlMapEntry)
+											.collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));	
+		productIds.stream()
+				.filter(id -> !productImgs.keySet().contains(id))
+				.forEach(id -> productImgs.put(id, NO_IMG_FOUND_URL));
+		
+		return productImgs;
+	}
+	
+
+	
+	
+	
+	private Map.Entry<Long, String> getProductCoverImageUrlMapEntry(Map.Entry<Long, List<ProductImagesEntity>> mapEntry){
+		String uri = Optional.ofNullable(mapEntry.getValue())
+							.map(List::stream)
+							.map(s -> s.sorted( comparing(ProductImagesEntity::getId)))
+							.flatMap(s -> s.findFirst())
+							.map(ProductImagesEntity::getUri)
+							.orElse(NO_IMG_FOUND_URL);
+				
+		return new AbstractMap.SimpleEntry<>(mapEntry.getKey(), uri);
 	}
 	
 	
@@ -831,11 +1027,52 @@ public class ProductImageServiceImpl implements ProductImageService {
 	}
 	
 	
+
+
+
+
+
+	@Override
+	public List<ProductImgDetailsDTO> getProductImgs(Long productId) throws BusinessException {
+		validateProductToFetchItsImgs(productId);
+		
+		return productImagesRepository.findByProductEntity_Id(productId)
+									.stream()
+									.map(this::toProductImgDetailsDTO)
+									.collect(Collectors.toList());
+	}
 	
 	
 	
-	private Integer compareByProductImageId(ProductImagesEntity img1, ProductImagesEntity img2) {
-		return Long.compare(img1.getId(), img2.getId());
+	
+	
+	private void validateProductToFetchItsImgs(Long productId) throws BusinessException{
+		if(!productRepository.existsById(productId)) {
+			throw new BusinessException("INVALID PARAM: product_id", ERR_NO_PRODUCT_EXISTS_WITH_ID, HttpStatus.NOT_ACCEPTABLE);
+		}
+		
+		validateUserCanModifyProduct(productId);
+	}
+
+
+
+
+
+
+	private ProductImgDetailsDTO toProductImgDetailsDTO(ProductImagesEntity entity) {
+		Long variantId = Optional.ofNullable(entity.getProductVariantsEntity())
+				.map(ProductVariantsEntity::getId)
+				.orElse(null);
+		
+		ProductImgDetailsDTO dto = new ProductImgDetailsDTO();
+		dto.setImageId(entity.getId());
+		dto.setPriority(entity.getPriority());
+		dto.setProductId(entity.getProductEntity().getId());
+		dto.setType(entity.getType());
+		dto.setUri(entity.getUri());		
+		dto.setVariantId(variantId);		
+		
+		return dto;
 	}
